+++ CHANGES.md
@@ -13,6 +13,8 @@
 
 - [ADD] `--client-cert` と `--client-key` でクライアント認証をできるようにする
     - @melpon
+- [ADD] Windows と Ubuntu で NVIDIA VIDEO CODEC SDK を使ったハードウェアデコーダに対応
+    - @melpon
 - [UPDATE] CLI11 を 2.2.0 に上げる
     - @voluntas
 - [FIX] Ubuntu 20.04 + H.264 + サイマルキャスト + --hw-mjpeg-decoder true で落ちるのを修正 (#221)